+++ CProcessing/src/Bosses/Boss.h
@@ -5,7 +5,7 @@ typedef struct
 {
 	CP_Vector Position; //Position vector of boss
 	int Health;			//Health of boss
-	int Size;			//Radius of boss for drawing
+	float Size;			//Radius of boss for drawing
 } Boss;
 
 Boss ArmorSlime; //currently only 1 boss
